#59 move ForgettablePayloads.Persistence implementation and abstractions to core ForgettablePayloads package
