+++ EventSourcing.ForgettablePayloads/ForgettablePayloads.Abstractions/Persistence/IForgettablePayloadStorageReader.cs
