+++ EventSourcing.ForgettablePayloads/ForgettablePayloads.Abstractions/Persistence/IForgettablePayloadStorageWriter.cs
