+++ EventSourcing.ForgettablePayloads/ForgettablePayloads/Persistence/ForgettablePayloadStorageReader.cs
