+++ EventSourcing.ForgettablePayloads/ForgettablePayloads/Persistence/ForgettablePayloadStorageWriter.cs
