+++ EventSourcing.ForgettablePayloads/ForgettablePayloads/Persistence/IForgettablePayloadStorageRepository.cs
