--- EventSourcing.ForgettablePayloads/Persistence/ForgettablePayloads.Persistence.Abstractions/ForgettablePayloads.Persistence.Abstractions.csproj
@@ -1,33 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <TreatWarningsAsErrors Condition=" '$(Configuration)' == 'Release' ">true</TreatWarningsAsErrors>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <RootNamespace>EventSourcing.ForgettablePayloads.Persistence</RootNamespace>
-        <AssemblyName>EventSourcing.ForgettablePayloads.Persistence.Abstractions</AssemblyName>
-    </PropertyGroup>
-
-    <!-- NuGet metadata -->
-    <PropertyGroup>
-        <Title>EventSourcing.ForgettablePayloads.Persistence.Abstractions</Title>
-        <Description>This package contains abstractions over EventSourcing.ForgettablePayloads.Persistence packages.</Description>
-        <Authors>Dawid Szmidka</Authors>
-        <PackageId>EventSourcing.ForgettablePayloads.Persistence.Abstractions</PackageId>
-        <Version>0.1.0-alpha</Version>
-        <PackageReleaseNotes>
-            - Initial version of EventSourcing.ForgettablePayloads.Persistence.Abstractions project
-        </PackageReleaseNotes>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageTags>EventSourcing;EventSourcing.ForgettablePayloads;EventSourcing.ForgettablePayloads.Persistence;EventSourcing.ForgettablePayloads.Persistence.Abstractions</PackageTags>
-        <IncludeSymbols>true</IncludeSymbols>
-        <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-        <IncludeSource>true</IncludeSource>
-        <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
-    </PropertyGroup>
-    
-    <ItemGroup>
-      <ProjectReference Include="..\..\ForgettablePayloads.Abstractions\ForgettablePayloads.Abstractions.csproj" />
-    </ItemGroup>
-
-</Project>
