--- EventSourcing.ForgettablePayloads/Persistence/ForgettablePayloads.Persistence.UnitTests/Extensions/ServiceCollectionExtensions_Should.cs
@@ -1,37 +1,0 @@
-﻿using EventSourcing.Extensions;
-using EventSourcing.ForgettablePayloads.Extensions;
-using EventSourcing.ForgettablePayloads.Persistence;
-using EventSourcing.ForgettablePayloads.Persistence.Extensions;
-using Microsoft.Extensions.DependencyInjection;
-using TestHelpers.Extensions;
-using Xunit;
-
-namespace ForgettablePayloads.Persistence.UnitTests.Extensions
-{
-    public class ServiceCollectionExtensions_Should
-    {
-        [Fact]
-        public void BuildWithoutErrors_When_AddingEventSourcingAndExternalDependenciesAndRequiredForgettablePayloadsPersistenceImplementation_Using_SeparateExtensionMethods()
-        {
-            var serviceCollection = new ServiceCollection();
-
-            serviceCollection
-                .AddEventSourcing()
-                .WithForgettablePayloads()
-                .WithForgettablePayloadsPersistence();
-            
-            // Required EventSourcing.ForgettablePayloads.Persistence implementations
-            serviceCollection.AddMock<IForgettablePayloadStorageRepository>();
-            
-            // External dependencies
-            serviceCollection
-                .AddEventSourcingSerializationMocks()
-                .AddEventSourcingPersistenceMocks()
-                .AddEventSourcingBusMocks()
-                .AddMicrosoftLoggerMock();
-
-            serviceCollection.BuildServiceProvider(
-                new ServiceProviderOptions { ValidateScopes = true, ValidateOnBuild = true });
-        }
-    }
-}
