--- EventSourcing.ForgettablePayloads/Persistence/ForgettablePayloads.Persistence.UnitTests/ForgettablePayloads.Persistence.UnitTests.csproj
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
-        <PackageReference Include="xunit" Version="2.4.1" />
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-            <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-        <PackageReference Include="coverlet.collector" Version="1.3.0">
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-            <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\..\Core\EventSourcing\EventSourcing.csproj" />
-      <ProjectReference Include="..\..\..\Helpers\TestHelpers\TestHelpers.csproj" />
-      <ProjectReference Include="..\..\ForgettablePayloads\ForgettablePayloads.csproj" />
-      <ProjectReference Include="..\ForgettablePayloads.Persistence\ForgettablePayloads.Persistence.csproj" />
-    </ItemGroup>
-
-</Project>
