--- EventSourcing.ForgettablePayloads/Persistence/ForgettablePayloads.Persistence/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("ForgettablePayloads.Persistence.UnitTests")]
-[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
