--- EventSourcing.ForgettablePayloads/Persistence/ForgettablePayloads.Persistence/ForgettablePayloads.Persistence.csproj
@@ -1,38 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <TreatWarningsAsErrors Condition=" '$(Configuration)' == 'Release' ">true</TreatWarningsAsErrors>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <RootNamespace>EventSourcing.ForgettablePayloads.Persistence</RootNamespace>
-        <AssemblyName>EventSourcing.ForgettablePayloads.Persistence</AssemblyName>
-    </PropertyGroup>
-
-    <!-- NuGet metadata -->
-    <PropertyGroup>
-        <Title>EventSourcing.ForgettablePayloads.Persistence</Title>
-        <Description>This package contains default implementation of EventSourcing.ForgettablePayloads.Persistence.Abstractions package.</Description>
-        <Authors>Dawid Szmidka</Authors>
-        <PackageId>EventSourcing.ForgettablePayloads.Persistence</PackageId>
-        <Version>0.1.0-alpha</Version>
-        <PackageReleaseNotes>
-            - Initial version of EventSourcing.ForgettablePayloads.Persistence project
-        </PackageReleaseNotes>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageTags>EventSourcing;EventSourcing.ForgettablePayloads;EventSourcing.ForgettablePayloads.Persistence</PackageTags>
-        <IncludeSymbols>true</IncludeSymbols>
-        <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-        <IncludeSource>true</IncludeSource>
-        <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
-    </PropertyGroup>
-    
-    <ItemGroup>
-      <ProjectReference Include="..\ForgettablePayloads.Persistence.Abstractions\ForgettablePayloads.Persistence.Abstractions.csproj" />
-    </ItemGroup>
-    
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />        
-    </ItemGroup>
-
-</Project>
