+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/ForgettablePayloads.Persistence.PostgreSql/ForgettablePayloads.Persistence.PostgreSql.csproj
@@ -27,13 +27,13 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\ForgettablePayloads.Persistence\ForgettablePayloads.Persistence.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
         <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />
         <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
         <PackageReference Include="Npgsql" Version="6.0.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\..\ForgettablePayloads\ForgettablePayloads.csproj" />
     </ItemGroup>
 
 </Project>
